Rename first attempt of plugins to event-subscribers
